1. Сделал дробовик с настройкой количества дробинок (параметр скрипта Enemy) 2. Сделал сохранения жизней игрока между сценами с настраиваемой генерацией (параметр скрипта Main) 3. Сделал новый вид врага Twister, поливающий свинцом заданный угловой сектор (тип и параметр скрипта Enemy)

+++ Assets/Resources/Scripts/Rocket.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class Rocket : MonoBehaviour
 {
-    public int damage; // урон
+    public float damage; // урон
     public float maxRange; // предельная дистанция полета
     public float speed; // скорость полета 
     public Vector3 startPoint; // начальная точка полета
